Merged substantial changes to SFT
During the month of Feb 2013, Vivek made substantial alterations to SFT
while Eric worked on the original merge of the modules version of MT3D
with SSPA's enhancements to MT3D.  This commit reflects Vivek's changes
(see Vivek's email dated Feb. 28, 2013).  Also, the MT3D-SSPA folder now
contains all of SSPA's updates as of 2/28/13.  This folder is used to
track the changes to SSPA's version of MT3D.